test: enhance validation tests for Optional and Union types in schema mapping
