+++ tests/llm/test_genai/test_invalid_schema.py
@@ -6,12 +6,13 @@ from google.genai import Client
 from pydantic import BaseModel
 from .util import models, modes
 from itertools import product
+from instructor.utils import map_to_gemini_function_schema
 
 
 @pytest.mark.parametrize("mode,model", product(modes, models))
 def test_nested(mode, model):
-    """Test that nested schemas raise appropriate error with Gemini."""
-    client = instructor.from_genai(Client(), mode=mode)
+    """Test that nested schemas are supported."""
+    client: instructor.Instructor = instructor.from_genai(Client(), mode=mode)
 
     class Address(BaseModel):
         street: str
@@ -21,17 +22,19 @@ def test_nested(mode, model):
         name: str
         address: Optional[Address] = None
 
-    with pytest.raises(ValueError, match="Gemini does not support Optional types"):
-        client.chat.completions.create(
-            model=model,
-            messages=[
-                {
-                    "role": "user",
-                    "content": "John loves to go gardenning with his friends",
-                }
-            ],
-            response_model=Person,
-        )
+    resp = client.chat.completions.create(
+        model=model,
+        messages=[
+            {
+                "role": "user",
+                "content": "John loves to go gardenning with his friends",
+            }
+        ],
+        response_model=Person,
+    )
+
+    assert resp.name == "John"  # type: ignore
+    assert resp.address is None  # type: ignore
 
 
 @pytest.mark.parametrize("mode,model", product(modes, models))
@@ -43,27 +46,182 @@ def test_union(mode, model):
         name: str
         id_value: Union[str, int]
 
-    with pytest.raises(ValueError, match="Gemini does not support Optional types"):
+    with pytest.raises(
+        ValueError,
+        match=r"Gemini does not support Union types \(except Optional\)\. Please change your function schema",
+    ):
         client.chat.completions.create(
             model=model,
             messages=[{"role": "user", "content": "User name is Alice with ID 12345"}],
             response_model=UserData,
-        )
+        )  # type: ignore
 
 
-@pytest.mark.parametrize("mode,model", product(modes, models))
-def test_optional(mode, model):
-    """Test that optional fields raise appropriate error with Gemini."""
-    client = instructor.from_genai(Client(), mode=mode)
+def test_optional_types_allowed():
+    """Test that Optional types are correctly mapped and don't throw errors."""
 
-    class Profile(BaseModel):
+    class User(BaseModel):
+        name: str
+        age: Optional[int] = None
+        email: Optional[str] = None
+
+    schema = User.model_json_schema()
+    # Should not raise an error
+    result = map_to_gemini_function_schema(schema)
+
+    assert result["properties"]["age"]["nullable"] is True
+    assert result["properties"]["email"]["nullable"] is True
+    assert result["required"] == ["name"]
+
+
+def test_union_types_rejected_schema():
+    """Test that Union types (not Optional) throw an error in schema mapping."""
+
+    class UserWithUnion(BaseModel):
+        name: str
+        value: Union[int, str]  # Should be rejected
+
+    schema = UserWithUnion.model_json_schema()
+
+    with pytest.raises(ValueError, match="Union types"):
+        map_to_gemini_function_schema(schema)
+
+
+@pytest.mark.parametrize(
+    "mode", [instructor.Mode.GENAI_STRUCTURED_OUTPUTS, instructor.Mode.GENAI_TOOLS]
+)
+def test_genai_api_call_with_different_types(mode):
+    """Test actual API call with genai SDK using different types."""
+
+    class UserProfile(BaseModel):
         name: str
         age: int
-        bio: Optional[str] = None
+        email: Optional[str] = None
+        is_premium: bool
+        score: float
 
-    with pytest.raises(ValueError, match="Gemini does not support Optional types"):
-        client.chat.completions.create(
-            model=model,
-            messages=[{"role": "user", "content": "Alice is 30 years old"}],
-            response_model=Profile,
-        )
+    client = instructor.from_provider("google/gemini-2.0-flash", mode=mode)
+
+    response = client.chat.completions.create(
+        messages=[
+            {
+                "role": "user",
+                "content": "Create a user profile for John Doe, 25 years old, premium user with score 85.5",
+            }
+        ],
+        response_model=UserProfile,
+    )
+
+    assert isinstance(response, UserProfile)
+    assert response.name == "John Doe"
+    assert response.email is None
+
+
+@pytest.mark.parametrize(
+    "mode", [instructor.Mode.GENAI_STRUCTURED_OUTPUTS, instructor.Mode.GENAI_TOOLS]
+)
+def test_genai_api_call_with_nested_models(mode):
+    """Test API call with nested models (multiple users)."""
+
+    class User(BaseModel):
+        name: str
+        age: int
+        department: Optional[str] = None
+
+    class UserList(BaseModel):
+        users: list[User]
+
+    client = instructor.from_provider("google/gemini-2.0-flash", mode=mode)
+
+    response = client.chat.completions.create(
+        messages=[
+            {
+                "role": "user",
+                "content": "Create a list of 3 employees: Alice (30, Engineering), Bob (25, Marketing), Charlie (35)",
+            }
+        ],
+        response_model=UserList,
+    )
+
+    assert isinstance(response, UserList)
+    assert len(response.users) == 3
+    assert {user.name for user in response.users} == {"Alice", "Bob", "Charlie"}
+    assert {user.age for user in response.users} == {25, 30, 35}
+    assert {user.department for user in response.users} == {
+        None,
+        "Engineering",
+        "Marketing",
+    }
+
+
+@pytest.mark.asyncio
+@pytest.mark.parametrize(
+    "mode", [instructor.Mode.GENAI_STRUCTURED_OUTPUTS, instructor.Mode.GENAI_TOOLS]
+)
+async def test_genai_api_call_with_different_types_async(mode):
+    """Test actual async API call with genai SDK using different types."""
+
+    class UserProfile(BaseModel):
+        name: str
+        age: int
+        email: Optional[str] = None
+        is_premium: bool
+        score: float
+
+    client = instructor.from_provider(
+        "google/gemini-2.0-flash", mode=mode, async_client=True
+    )
+
+    response = await client.chat.completions.create(
+        messages=[
+            {
+                "role": "user",
+                "content": "Create a user profile for John Doe, 25 years old, premium user with score 85.5",
+            }
+        ],
+        response_model=UserProfile,
+    )
+
+    assert isinstance(response, UserProfile)
+    assert response.name == "John Doe"
+    assert response.email is None
+
+
+@pytest.mark.asyncio
+@pytest.mark.parametrize(
+    "mode", [instructor.Mode.GENAI_STRUCTURED_OUTPUTS, instructor.Mode.GENAI_TOOLS]
+)
+async def test_genai_api_call_with_nested_models_async(mode):
+    """Test async API call with nested models (multiple users)."""
+
+    class User(BaseModel):
+        name: str
+        age: int
+        department: Optional[str] = None
+
+    class UserList(BaseModel):
+        users: list[User]
+
+    client = instructor.from_provider(
+        "google/gemini-2.0-flash", mode=mode, async_client=True
+    )
+
+    response = await client.chat.completions.create(
+        messages=[
+            {
+                "role": "user",
+                "content": "Create a list of 3 employees: Alice (30, Engineering), Bob (25, Marketing), Charlie (35)",
+            }
+        ],
+        response_model=UserList,
+    )
+
+    assert isinstance(response, UserList)
+    assert len(response.users) == 3
+    assert {user.name for user in response.users} == {"Alice", "Bob", "Charlie"}
+    assert {user.age for user in response.users} == {25, 30, 35}
+    assert {user.department for user in response.users} == {
+        None,
+        "Engineering",
+        "Marketing",
+    }
